+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Pending
 
+### Update:
+- feat: add functions to send assets to contract address without relying on Stellar RPC. ([#677](https://github.com/stellar/java-stellar-sdk/pull/677))
+
 ## 1.1.0
 
 In this release, we have significantly simplified the process of using the Java SDK to interact with Soroban contracts. 
@@ -9,7 +12,7 @@ You can now use stellar-contract-bindings to generate Java bindings for contract
 Please visit [the GitHub repository](https://github.com/lightsail-network/stellar-contract-bindings) of stellar-contract-bindings for more information.
 
 ### Update
-- feat: add `ContractClient` to simplify contract calls. ([#675](https://github.com/stellar/java-stellar-sdk/pull/675))
+- feat: add `ContractClient` to simplify contract calls. ([#677](https://github.com/stellar/java-stellar-sdk/pull/675))
     
 ## 1.0.0